Mind bending higher-order classes.

+++ get_stuff.sh
@@ -10,22 +10,28 @@ checkret () {
 
 if [ ! -d python-ovr ]; then
     hg clone https://bitbucket.org/duangle/python-ovr
+else
+    cd python-ovr
+    hg pull -u
+    cd ..
 fi
 
-echo "==== Copying OculusSDK dir to python-ovr dir and building... ===="
-cp -rf ./OculusSDK ./python-ovr/ovr-src
-cd python-ovr/ovr-src/
+if [ -d OculusSDK && ! -f oculus_lock]; then
+    echo "==== Copying OculusSDK dir to python-ovr dir and building... ===="
+    cp -rf ./OculusSDK ./python-ovr/ovr-src
+    cd python-ovr/ovr-src/
 
-./ConfigurePermissionsAndPackages.sh
-checkret
+    ./ConfigurePermissionsAndPackages.sh
+    checkret
 
-make -j
-checkret
+    make -j
+    checkret
 
-cd ..
+    cd ..
 
-python setup.py develop
-checkret
+    python setup.py develop
+    checkret
+fi
 
 echo "==== Hopefully everything went OK. ==== "
